[IMP] account_{external,avatax}_tax,l10n_br_edi: EDI for vendor bills
This does both tax calculation and EDI for vendor bills and vendor
bill refunds. It's needed because in certain cases the customer will
pay a specific tax on bills that is not directly charged by the
vendor.

Luckily, not many changes are needed to adapt the tax calculation and
EDI code for invoices to work on bills as well. The most important
ones are changing the operation type and allowing to use the
account.move.send wizard to work with bills.

Because this needs to change what `move_type`s an external tax
calculator acts on, we take the opportunity to absorb those conditions
into the `is_tax_computed_externally` field. It simplifies the logic a
bit, and makes it easier to change based on the needs of the tax
calculator. It means that every tax calculator is now responsible for
marking the types of records they act on in a
`_compute_is_tax_computed_externally` override. It's necessary because
the move_type domain used to be part of the view in *external_tax*,
it's not possible to add that check in the default
`_compute_is_tax_computed_externally` because then every sale document
would be considered eligible for external taxes.

task-4608570

closes odoo/enterprise#83476

Related: odoo/odoo#208175
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>