[IMP] helpdesk_timesheet: improve list view in ticket form view
This commit changes the order uses inside the list view to display the
most recent timesheets first. Also, when the user would like to create
a new timesheet in that list view, the new row will be on the top of
the list since in the most of the cases, the user will create a
timesheet for today.

task-4637865

closes odoo/enterprise#81153

Related: odoo/odoo#201189
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>