[FIX] approvals: reset list of approvers on category change
The list of approvers is computed based on the selected category for
the approval request. Before the fix, the list would keep previous
approvers when changing the category. This could lead to issues where
users could add themselves as approvers on unauthorized categories.

Steps to reproduce:
- Create an approval request category (c1) with yourself as approver.
- Create another approval request category (c2) with someone else as approver.
- Create a request with category c1 and change to category c2.
- You are now an approver of the approval request.

This commit fixes the issue by reseting the list of approvers to only
take the ones configured on the category.

task-4637083

closes odoo/enterprise#82406

X-original-commit: 328033c76e015fdcf570063b6cb33d7469be28f1
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>