[IMP] l10n_in_reports: improve GST token generation for shared GST numbers
Before This Commit:
If a company attempted to generate a GST token using a GST number already
assigned to another company with a valid token, the system returned an ambiguous
error:  {'message': 'API access is not available or user expiry duration is less
than or equal to auth token expiry duration', 'error_cd': 'AUTH4037'}. This
caused confusion for users, as the message did not clearly indicate the issue.

After This Commit:
A new function has been introduced to check whether another company has the same
GST number and a valid token before sending the request. If such a case is
detected, the system now displays a warning, recommending the use of a Tax Unit.

task-4636041

closes odoo/enterprise#84874

X-original-commit: 0dbec10049a73cf855abfc3c6fa2d33e8961982e
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Harsh Trivedi (hatr) <hatr@odoo.com>