[IMP] documents: improve the documents list view
Purpose
=======
Opening the preview / navigating in folders is now done
by clicking on the (mime) type icon, leaving more room to
only focus on the record, as will be used in a later commit
to update the details panel.

Show the file size like we do in the details panel.

Task-4596337

Part-of: odoo/enterprise#80138
Related: odoo/odoo#201267
Related: odoo/upgrade#7475
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>