[FIX] timesheet_grid: hide `is_timer_running` column in the list view
Before this commit, the `is_timer_running` field defined timesheet model
is displayed in the list but the column is always empty.

This commit makes sure the column is also hidden in the list view since
no data is displayed. The field is in fact just a technical field and
not really useful for the user.

closes odoo/enterprise#81155

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>