[FIX] l10n_*: remove `countries` from the manifest
After https://github.com/odoo/odoo/pull/144586, the modules
other than base localization module should not have
the `countries` key defined in their manifest.

The issue happens because `countries` key works similar
to auto-install and it can lead to unwanted installation
of the module.

As there was no check implemented, for some modules this
key was used accidentally and caused issues with auto-install.

This commit removes this field from the manifest of the
modules that could have this problem to avoid the issue.

Also in https://github.com/odoo/odoo/pull/201526 a linting
test is proposed to avoid having the same issue in the future.

closes odoo/enterprise#83096

X-original-commit: de2a239905486354d6c4069352fa13c17bebe92f
Related: odoo/odoo#205273
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>