[IMP] web_studio: statusbar buttons display first, dropdown rest
This commit is a follow-up to the one that adapts the statusbar buttons
layout on smaller screen to make them more discoverable than moving them
all to the CogMenu.

In a nutshell, the first button is displayed like on bigger screens, but
the other ones (if any) are folded into a dropdown.

task-4586651

closes odoo/enterprise#81233

Related: odoo/odoo#198621
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>