[FIX] mrp_workorder,quality: missing compute depends
This commit adds the missing depends on `_compute_show_failure_location`
to make in work in onchange context + make sure that only operation and
test_type have an influence on it. The manufacturing picking type should
not be taken into account to hide the failure location.

Task: 4632935
X-original-commit: ffcd46e1a677a75a6971aa42edf0315c4bf67ace
Part-of: odoo/enterprise#88088
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>