[IMP] l10n_be_reports: Fiscal position from wrong company
Ensure that the correct fiscal position is used in
test_generate_ec_sales_list_activity.

Previously, the search for a fiscal position did not
filter by the company where the invoice was created.
As a result, it returned the latest "Intra-Community"
fiscal position in the database for the allowed companies,
potentially leading to incorrect assignments.

This commit adds a call to _check_company_domain in the
search domain, ensuring that the fiscal position is
correctly retrieved for the intended company.

runbot-145700

closes odoo/enterprise#81294

X-original-commit: a1fdfe284ac2a7f999c10fbb4578880d18e02baa
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>