[IMP] stock_barcode: shift 'Put in Pack' barcodes in demo sheet
The 'Put in Pack' barcodes in the Code 128 section of the barcode
demo sheet were moved from the bottom of the first page to the top
of the second page by painstakingly modifying the hardcoded PostScript
coordinates. This way all barcodes remain grouped under their
respective headers.

Note: The `%!PS` string is simply the Postscript file signature magic
string. It has no functional effect on the resulting file, but it's good
practice to include it for compatibility's sake.

Task ID: 3981936

closes odoo/enterprise#81265

Related: odoo/odoo#196246
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>