[FIX] quality{,_control}: split stock move with failed location
In case of 'quantity' control per quality point. If a complete stock
move line is sent to a failure location. The stock move was not split.
We could have a stock move line going to a location that is not child of
the location of the corresponding move. This can lead issues in case of
push_rules or stock valuation.

closes odoo/enterprise#91474

Task: 4575193
X-original-commit: 4b6b4051ca806372aa4d040a0094240df48a8a02
Related: odoo/odoo#221469
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>