[IMP] spreadsheet_edition: prevent traceback when a spreadsheet cannot be accessed
Before this commit, if a user attempted to open a spreadsheet they lacked access to
or that did not exist, they would get a traceback due to improper error handling.
This commit resolves the issue by handling server errors when fetching the spreadsheet
and redirecting the user to the home page.

Steps to reproduce:
- Create a spreadsheet with Mitchell Admin.
- Copy a hyperlink.
- In another browser, log in as Marc Demo and paste the hyperlink.
- You get a traceback and you don't understand what's going on.

task-4551255

closes odoo/enterprise#86328

X-original-commit: 96e6308b402a0fcb7d960086451fe806e3f538c9
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Mehdi Rachico (mera) <mera@odoo.com>