[IMP] (documents_)hr_payroll: improve handling of payslips for employee without user
If a employee has no user, there is no 'My Drive' linked to the employee and
the payslip cannot be store in someone's My Drive.

The payslip is instead placed into a new "Workers Payslips" folder that is only
accessible by the super admin in the root of company documents. This folder is
only created at payslip generation if the employee has no user. Otherwise, if
there aren't no employee without user, the folder is never created. The goal
is to avoid doing noise on company root by placing all the payslip there, but
the payslip still need to be placed somewhere in order to be accessible for the
employee in the document app via the share link.

That document link is now directly provided in the mail noticing the employee
that a new payslip is available. As the document is not accessible by anyone,
except super admin, the employee still have to receive the payslip.

Note: We were first thinking of ensuring that the employee's HR documents would
be moved to 'follow the employee' whenever the user would be changed on the
employee. However, this is a niche use case, were we wouldn't know if the
change would be on purpose or by mistake. Also, documents are generated for
the user behind the employee, not for the employee itself. So changing the user
on the employee should not make the former user lose its own documents.

Task-4309991

Part-of: odoo/enterprise#80399
Related: odoo/odoo#203729
Related: odoo/upgrade#7532
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>