[FIX] delivery_fedex_rest:multi package return label warning message
To reproduce:
- Install delivery_fedex_rest and enter API credentials for FedEx US,
  also turn on return label generation.
- Enable packages in inventory settings.
- New delivery to Azure Interior, 1x product 5555, save and Put in Pack
  (FedEx Box 1kg).
- Add another line with 1x product 5555, save and Put in Pack again
  (FedEx Box 1kg).
- Validate the picking.

Current behaviour:
Crashes when trying to create the return label with nondescript error
coming from FedEx.

Expected behaviour:
Ideally, a correctly validated picking with generated return labels.
However, this appears to currently not be supported by the FedEx REST
API... (multi-package single-shot return shipments of print label type)
We compromise and will create the outgoing shipment as usual, but not
the return shipment in case of multi-package shipments.
A helpful message informing the user of this limitation will be put in
the chatter of the picking instead of the generated return label in this
specific case.

opw-4556415

closes odoo/enterprise#81437

X-original-commit: 1140376cf5a09e94f348ad56a17573f6420b888a
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>