[IMP] iot: always open websocket channel
We used to open the websocket channel between iot box and db only when a report was set, to opening a useless connection.
In order to simplify the code and reduce complexity when using ws for anything else than printing a report, we removed
this logic.
Note that we only use one ws channel for all iot boxes on the same company, making the load very small.

closes odoo/enterprise#81332

Related: odoo/upgrade#7383
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>