[FIX] account_accountant: reco wizard: exchange rate rounding test
In the related community commit we improved the reconciliation
behavior with regards to exchange rate and rounding issues.
See there for more details.

For one of the tests there we manually created the write-off line with the
correct amount.
Here we add a test to check that the wizard computes the right amounts
for the write-off line / entry.

opw-4438542

closes odoo/enterprise#84875

X-original-commit: a8367938eb8b3a197926e1fa4eda6209a002d842
Related: odoo/odoo#208798
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>