[FIX] account_accountant: reco wizard: show company amount
The reconciliation wizard may create a write-off line / entry.
Roughly speaking this happens in case
- the amount of the input lines does not sum to 0
- and the "Allow partials" option is not ticked

The write-off line can be in foreign currency.
Currently the wizard only displays the amount in foreign currency but
not the amount in company currency.
So currently the company currency amount can not be checked before reconciling.

After this commit the wizard also displays the amount in company currency.

opw-4623933

X-original-commit: 65fbb9e89fedd32a2826418c2f8bff531bbca71e
Part-of: odoo/enterprise#84875
Related: odoo/odoo#208798
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>