[IMP] iot: add access groups
Before this commit, the IoT app had no access
control other than needing to be an admin. This
meant that the only way to manage IoT was to be
an admin for all of Odoo.

After this commit, IoT user and IoT admin groups
have been added. IoT users can access the IoT app
but not modify anything, whereas IoT admins have
full access.

In order to add these groups, the IoT module had
to be moved to one of the predefined categories,
in this case 'Administration'.

task-4643403

closes odoo/enterprise#81362

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>