[FIX] stock_barcode_mrp: handle scan package from type in MO
**Current behavior:**
Scanning a package type in an MO in barcode while having a raw
move line (component line) selected will result in a traceback.

**Expected behavior:**
Put the comp line in a new pack of that package type.

**Steps to reproduce:**
1. Make a packge type

2. Make a new MO -> open it in barcode

3. Scan some component, select the new line

4. Scan the barcode of the package type -> traceback

**Cause of the issue:**
We end up calling `action_put_in_pack` when `this.resModel` is
`MrpProduction` (which does not define this method).

**Fix:**
Define the method for the barcode MRP model to display a
notification to, rather, scan a product (scanning a package here
doesn't make sense).

opw-4564834

closes odoo/enterprise#87877

X-original-commit: 902d53d1ad13eeac684579cd2dac9a0dce8d7f21
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>