[IMP] stock_barcode: removes malformed CSS animation
This commit removes the malformed (and thus unapplied)
`highlighting-flash-primary` CSS animation.

In a nutshell, the animation as written in SCSS doesn't interpret the
`theme-color("primary")` call and write it as-is in the generated CSS...
which, in return, is an unknown CSS value.

Note: this commit also removes the `highlighting-flash-white` animation
as it isn't used.

closes odoo/enterprise#81592

Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>