[IMP] planning: avoid assign archived shift when duplicating shift
Before this commit, when the user duplicates a shift with resource
archived, the new shift will still contain the same resource even if the
resource no longer works for the company.

This commit does not assign the archived resource on the new shift
duplicated (when the user duplicates a shift with archived resource).
However, if the resource is given as default value during the
duplication or the duplication is used by the planning split tool feature
than the resource will not be checked if it is archived or not.

task-4419770

closes odoo/enterprise#76683

Related: odoo/odoo#192848
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>