[IMP] {website_}appointment{_crm}: automatically assign salespersons to visitors
Each time a visitor creates an appointment, a lead is created too. Therefore,
the pipe can hold a lot of leads very quickly.

This commit aims to reuse leads when it is possible instead of create new ones.
For this, if the appointment type has a random staff user selection, the last
ongoing lead of the visitor is used to chose the staff user of the event. Then
the new appointment is linked to the visitor's last ongoing lead with the
selected staff user, if it is still available, otherwise a new lead is created.

task-4247031

Part-of: odoo/enterprise#77579
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>