[FIX] stock_barcode: scan barcode in form view
Issue
=====
Scanning a product while beeing in the stock move line Barcode form view
or in the quand Barcode form view can cause a traceback and the
impossibility to exit a Barcode operation.

How to reproduce
================
1. Create two products with a barcode;
2. Open the Barcode app and create an operation;
3. Scan a product then click on the edit button to open the form view;
4. Update the product quantity field;
5. While still in the form view, scan another barcode -> You can see the
   quantity was reset;
6. Update again the quantity and save;
7. Try to exit the operation -> Bim badaboum, traceback 💥!

Cause of the issue
==================
When a barcode is scanned, the app doesn't check where is the current
state and process the barcode anyway which lead to strange behavior and
inconsistencies between the current lines and the lines to save.

Solution
========
Disable the scan while somewhere else that in the barcode line view.

OPW-4567723

closes odoo/enterprise#85191

X-original-commit: a8df4053cb08099a8711c5179ce30cdc8daab701
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>