[FIX] l10n_mx_edi: fix withholding tax rounding issue
Steps to reproduce:
- Install "l10n_mx_edi" and switch to a Mexican company
- In Accounting > Configuration > Taxes activate the tax "1.25 WH"
- Create an invoice with one line with a price of $50 and the tax "1.25 WH"
- Go to the customer of the invoice
- Sale & Purchase > Fiscal information and activate "No Tax Breakdown"
- Confirm the invoice and Send it to CFDI
- Traceback

Cause:
This [commit](https://github.com/odoo/enterprise/pull/75381/commits/3f2e9437439fd6f987a13816473da2bd5a26953d) added a part in the code to prevent rounding error when specific discounts are applied which makes the base have a 5 as third digit.

When having a withholding tax and selecting "No Tax Breakdown" on the partner, the base amount will be decreased by the amount of the tax. If this amount has a 5 as third digit, a difference is detected here:
`delta_discount = currency.round(sum(x['importe'] for x in line_values_list)) - cfdi_values['subtotal']`
because:
- `x['importe']` is 49.375 which rounds up to 49.38
- `cfdi_values['subtotal']` is 49.37 because the tax was rounded up from 0.625 to 0.63
The code will then try to find lines with discounts but there are none so it raises an error at the `max()` function which is given nothing.

Solution:
Adding `default=None` to `max()` so it doesn't crash.

But the rounding issue still needs to be fixed.
The fix is to select a line with more than 2 digits and to remove the difference with its rounded value.
This ensures that the resulting value stay in the same 2-digit range, preventing error CFDI40167, and that the sum of "importe" is rounded down.

opw-4528766

closes odoo/enterprise#82669

X-original-commit: b723e23bec0f8748014b950ae8af7c5622248b39
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>