[REF] spreadsheet_edition, documents_spreadsheet: Improve perf of global filters tests
This commit has two main goals:
- Reduce memory usage/time of global filters tests
- Move global filters tests to their module (from documents_spreadsheet
to spreadsheet_edition)

The tests have been refactored to use the minimum setup possible.
Before this commit, tests were using the full setup with a web client,
with a pivot view and then a click on "Insert in spreadsheet", then the
rendering of the spreadsheet, then the opening of the global filters
side panel. And only after that the test can begin.

Now, we create a spreadsheet model and we mount only the components that
are tests (the global filters side panel and the global filter editor).

The tests are now faster and use less memory.

| On runbot                                      | Before | After |
|------------------------------------------------|--------|-------|
| Memory consumption of all JS tests, at the end | 880mb  | 715mb |
| Time to run this suite                         | 40s    | 9s    |

closes odoo/enterprise#81605

Task: 4653513
Related: odoo/odoo#202112
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>