[FIX] sale_commission_subscription: convert achieved amount
The MRR log amount was stored in the currency of the SO.
Before this commit, that amount was assumed to be in the currency of
the plan before being converted into the currency of the current
company.

This commit reuse the mechanism of the MRR report to convert the amount
from the SO currency to the current company currency directly.

Taskid: 4652771
X-original-commit: 133874f491a92229b4de6fb1f6b0d94b7b5eb3c9
Part-of: odoo/enterprise#82508
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>